// frontend/src/lib/api/generated.ts
export interface components {
    schemas: {
// ...
        ProfileInput: {
            fullName: string;
            specialization: string;
            timezone: components["schemas"]["TimeZone"];
            institute?: string;
            course?: number;
            department?: string;
            company?: string;
            position?: string;
            avatarMediaId?: components["schemas"]["Uuid"];
        };
        Profile: {
            id: components["schemas"]["Uuid"];
            accountId: components["schemas"]["Uuid"];
            formalRole: components["schemas"]["FormalRole"];
            publicationState: components["schemas"]["PublicationState"];
            published?: components["schemas"]["ProfileInput"];
            pending?: components["schemas"]["ProfileInput"];
            editLocked: boolean;
            moderation?: components["schemas"]["ModerationStatus"];
            rowVersion: number;
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            updatedAt: string;
        };
// ...
        Resume: {
            id: components["schemas"]["Uuid"];
            slot: number;
            primary: boolean;
            searchVisible: boolean;
            publicationState: components["schemas"]["PublicationState"];
            published?: components["schemas"]["ResumeInput"];
            pending?: components["schemas"]["ResumeInput"];
            editLocked: boolean;
            moderation?: components["schemas"]["ModerationStatus"];
            rowVersion: number;
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            updatedAt: string;
        };
// ...
    };
// ...
}